refactor: creating metrics helper in train nodejs app
