+++ other/train-to-cloud-city/devices/rfid/utils/firestoreHelpers.js
@@ -15,20 +15,7 @@
 const { firebase, db, app, firestore } = require("./firebase.js");
 const { StringDecoder } = require("node:string_decoder");
 const { getMotor } = require("./train.js");
-const { PubSub } = require("@google-cloud/pubsub");
-
-const pubSubClient = new PubSub();
-
-async function publishMessage(topic, data) {
-  const dataBuffer = Buffer.from(JSON.stringify(data));
-  try {
-    const messageId = await pubSubClient.topic(topic).publishMessage({ data: dataBuffer });
-    console.log(`Message ${messageId} published.`);
-  } catch (error) {
-    console.error(`Received error while publishing: ${error.message}`);
-    process.exitCode = 1;
-  }
-}
+const { queueMessageToPublish } = require("./metrics.js");
 
 /**
  * setMissionPattern
@@ -53,11 +40,10 @@ async function setMissionPattern(chunk, reader) {
         { merge: false },
       );
 
-      /*await publishMessage("mission-selected", {
+      queueMessageToPublish("mission-selected", {
         chunk,
-        timestamp: Date.now(),
         pattern_slug: matchingTag?.pattern_slug,
-      });*/
+      });
 
       console.log(
         `Mission has been read: ${JSON.stringify(matchingTag?.pattern_slug)}. Waiting for event input trigger.`,
@@ -235,10 +221,9 @@ async function updateLocation(location) {
   const ref = db.collection("global").doc("train");
   try {
     await ref.update({ actual_location: location }, { merge: true });
-    /*await publishMessage("location-updated", {
-      location,
-      timestamp: Date.now(),
-    });*/
+    
+    queueMessageToPublish("location-updated", { location });
+    
     console.log(`Passed checkpoint ${location}`);
   } catch (error) {
     console.error(error);
