+++ other/train-to-cloud-city/devices/rfid/utils/metrics.js
@@ -1,0 +1,38 @@
+const { PubSub } = require("@google-cloud/pubsub");
+
+const pubSubClient = new PubSub();
+
+let queuedMetricsToPublish = [];
+
+/**
+ * queueMessageToPublish
+ * ---------------------------
+ * Queue up metrics to publish while game is going on
+ */ 
+function queueMessageToPublish(topic, data) {
+  const dataBuffer = Buffer.from(JSON.stringify(data));
+  queuedMetricsToPublish.push({ topic, data: dataBuffer, timestamp: Date.now() });
+}
+
+/**
+ * publishQueuedMessages
+ * ---------------------------
+ * As long as there are items in the queue to publish
+ * continue pushing items up to google-cloud pubsub
+ */ 
+(async function publishQueuedMessages() {
+  while(queuedMetricsToPublish?.length) { 
+    queuedMetricsToPublish?.forEach((metric) => {
+      try {
+        const messageId = await pubSubClient.topic(metric?.topic).publishMessage({ data: metric?.data, timestamp: metric?.timestamp });
+        console.log(`Message ${messageId} published.`);
+      } catch (error) {
+        console.error(`Received error while publishing: ${error.message}`);
+        process.exitCode = 1;
+      }
+    });
+  }
+})();
+
+module.exports = { queueMessageToPublish };
+
